Add blog page and fix SPA routing
The app were not getting injected when refreshing
in a nested route, because of the history API. To
fix this I've to set --history-api-fallback opt
to the webpack-dev-server execution and set the
public path to "/".

Also, a 404 HTML was added to fix this problem in
Github Pages. This file has a script which
replaces the route params with a query string and
redirect to the root. Then, the index is loaded
and the app injected, setting the proper params
to make it to work as expected.