Simplify module instantiation and extend blog
Many changes were made in this commit, but the main focus was the
improvement of the module registering and instantiation. The old
decorator as well as the Store class were removed, just because they
made everything more complicated and opaque. A main Vue component is
now the entry point of the module.

The blog module was also improved:
* A Github token was added to increase the messages limit.
* A component to manage the transitions between routes was added.
* The blog pages were improved and some styles were applied.